refactor: 프로필모달 컴포넌트 src/shared로 이관 및 import문 수정
